commit #2
Player and enemy class are renamed . Move method is changed to abstract and overrode for both player and enemy. Enemy doesn't require rigidbody so transform and vector logic is implemented.
