+++ Assets/Scripts/Character.cs
@@ -10,7 +10,6 @@ public class Character : MonoBehaviour
     public int dexterity = 5;
     public string weapon; //we can later create a weapon class and use that instead of string
     protected int currentHealth;
-    protected Rigidbody2D rb;
     protected virtual void Start()
     {
         rb = GetComponent<Rigidbody2D>();
@@ -19,13 +18,6 @@ public class Character : MonoBehaviour
     protected virtual void Update()
     {
         Move();
-    }
-    protected virtual void Move()
-    {
-        float horizontalInput = Input.GetAxis("Horizontal");
-        float verticalInput = Input.GetAxis("Vertical");
-        Vector2 movement = new Vector2(horizontalInput, verticalInput).normalized;
-        rb.velocity = movement * moveSpeed;
     }
     public virtual void TakeDamage(int damage)
     {
@@ -39,4 +31,5 @@ public class Character : MonoBehaviour
     {
         Destroy(gameObject);
     }
+    public abstract void Move();
 }
